Refactored routing functionality from server.js to router.js and added a logger called morgan
