+++ package.json
@@ -27,6 +27,7 @@
     "body-parser": "1.15.x",
     "ejs": "^2.4.2",
     "express": "4.x.x",
-    "express-validator": "2.20.x"
+    "express-validator": "2.20.x",
+	"morgan": "1.7.0"
   }
 }
