+++ public/signup.html
@@ -3,13 +3,14 @@
 	<head>
 		<title>Car Detail Exchange - Sign Up</title>
 
+		<!-- in server.js, the express.static points to public/assets: that's the "current directory" -->
 		<!-- <link rel="stylesheet" href="css/bootstrap.min.css"> -->
 		<link rel="stylesheet" href="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css">
 		<link rel="stylesheet" href="css/style.css">
-		<script src="assets/scripts/jquery-3.0.0.min.js" defer></script>
+		<script src="scripts/jquery-3.0.0.min.js" defer></script>
 		<script src="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script>
 		<!-- <script src="assets/scripts/bootstrap.min.js" defer></script> -->
-		<script src="assets/scripts/scripts.js" defer></script>
+		<script src="scripts/scripts.js" defer></script>
 	</head>
 	
 	<body>
